Moving stuff into core-android module
