+++ client-data/build.gradle
@@ -24,8 +24,6 @@ dependencies {
 
   api 'com.google.dagger:dagger:2.21'
 
-  api 'androidx.lifecycle:lifecycle-extensions:2.0.0'
-
   api 'com.squareup.retrofit2:retrofit:2.5.0'
   api 'com.squareup.retrofit2:converter-gson:2.5.0'
   api 'com.squareup.retrofit2:adapter-rxjava2:2.5.0'
