+++ client/src/main/java/com/jraska/github/client/NotificationSetup.kt
@@ -5,6 +5,7 @@ import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.content.Context
 import android.os.Build
+import com.jraska.github.client.core.android.OnAppCreate
 
 import javax.inject.Inject
