+++ client/src/main/java/com/jraska/github/client/SetupLogging.kt
@@ -3,7 +3,8 @@ package com.jraska.github.client
 import android.app.Application
 import com.jraska.console.timber.ConsoleTree
 import com.jraska.github.client.core.android.AnalyticsLoggingTree
-import com.jraska.github.client.logging.ErrorReportTree
+import com.jraska.github.client.core.android.OnAppCreate
+import com.jraska.github.client.core.android.logging.ErrorReportTree
 import timber.log.Timber
 import javax.inject.Inject
 import javax.inject.Provider
