+++ client/src/main/java/com/jraska/github/client/settings/SettingsActivity.kt
@@ -7,7 +7,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.airbnb.epoxy.SimpleEpoxyAdapter
 import com.jraska.github.client.R
 import com.jraska.github.client.ui.BaseActivity
-import com.jraska.github.client.viewModel
+import com.jraska.github.client.core.android.viewModel
 import kotlinx.android.synthetic.main.content_settings.*
 
 class SettingsActivity : BaseActivity(), PurchaseReportModel.PurchaseListener {
