+++ client/src/main/java/com/jraska/github/client/ui/UriHandlerActivity.kt
@@ -3,7 +3,7 @@ package com.jraska.github.client.ui
 import android.os.Bundle
 import com.jraska.github.client.core.android.UriHandlerViewModel
 import com.jraska.github.client.inputUrl
-import com.jraska.github.client.viewModel
+import com.jraska.github.client.core.android.viewModel
 
 class UriHandlerActivity : BaseActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
