+++ client/src/main/java/com/jraska/github/client/ui/UserDetailActivity.kt
@@ -12,7 +12,7 @@ import com.jraska.github.client.R
 import com.jraska.github.client.users.RepoHeader
 import com.jraska.github.client.users.UserDetail
 import com.jraska.github.client.users.UserDetailViewModel
-import com.jraska.github.client.viewModel
+import com.jraska.github.client.core.android.viewModel
 import kotlinx.android.synthetic.main.activity_user_detail.*
 import kotlinx.android.synthetic.main.content_user_detail.*
