+++ core-android/build.gradle
@@ -19,6 +19,8 @@ dependencies {
   api 'androidx.appcompat:appcompat:1.1.0-alpha04'
   api 'androidx.annotation:annotation:1.0.2'
 
+  api 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+
   api 'com.jakewharton.timber:timber:4.7.1'
 
   testImplementation 'junit:junit:4.12'
